Refactor GlobalSearch
