+++ zipkin-lens/src/containers/GlobalSearch/GlobalSearchContainer.js
@@ -12,57 +12,18 @@
  * the License.
  */
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router';
 
-import GlobalSearch from '../../components/GlobalSearch';
-import { fetchRemoteServices } from '../../actions/remote-services-action';
-import { fetchSpans } from '../../actions/spans-action';
-import { fetchServices } from '../../actions/services-action';
+import GlobalSearch from '../../components/GlobalSearch/GlobalSearch';
 import { fetchTraces } from '../../actions/traces-action';
-import { fetchAutocompleteKeys } from '../../actions/autocomplete-keys-action';
-import { fetchAutocompleteValues } from '../../actions/autocomplete-values-action';
-import {
-  setLookbackCondition,
-  setLimitCondition,
-  addCondition,
-  deleteCondition,
-  changeConditionKey,
-  changeConditionValue,
-} from '../../actions/global-search-action';
 
 const mapStateToProps = state => ({
-  services: state.services.services,
-  isLoadingServices: state.services.isLoading,
-  remoteServices: state.remoteServices.remoteServices,
-  isLoadingRemoteServices: state.remoteServices.isLoading,
-  spans: state.spans.spans,
-  isLoadingSpans: state.spans.isLoading,
   conditions: state.globalSearch.conditions,
   lookbackCondition: state.globalSearch.lookbackCondition,
   limitCondition: state.globalSearch.limitCondition,
-  autocompleteKeys: state.autocompleteKeys.autocompleteKeys,
-  isLoadingAutocompleteKeys: state.autocompleteKeys.isLoading,
-  autocompleteValues: state.autocompleteValues.autocompleteValues,
-  isLoadingAutocompleteValues: state.autocompleteValues.isLoading,
 });
 
 const mapDispatchToProps = dispatch => ({
-  fetchServices: () => dispatch(fetchServices()),
-  fetchRemoteServices: serviceName => dispatch(fetchRemoteServices(serviceName)),
-  fetchSpans: serviceName => dispatch(fetchSpans(serviceName)),
   fetchTraces: params => dispatch(fetchTraces(params)),
-  fetchAutocompleteKeys: () => dispatch(fetchAutocompleteKeys()),
-  fetchAutocompleteValues: autocompleteKey => dispatch(fetchAutocompleteValues(autocompleteKey)),
-  setLookbackCondition: lookbackCondition => dispatch(setLookbackCondition(lookbackCondition)),
-  setLimitCondition: limitCondition => dispatch(setLimitCondition(limitCondition)),
-  addCondition: condition => dispatch(addCondition(condition)),
-  deleteCondition: index => dispatch(deleteCondition(index)),
-  changeConditionKey: (index, conditionKey) => {
-    dispatch(changeConditionKey(index, conditionKey));
-  },
-  changeConditionValue: (index, conditionValue) => {
-    dispatch(changeConditionValue(index, conditionValue));
-  },
 });
 
 const GlobalSearchContainer = connect(
@@ -70,4 +31,4 @@ const GlobalSearchContainer = connect(
   mapDispatchToProps,
 )(GlobalSearch);
 
-export default withRouter(GlobalSearchContainer);
+export default GlobalSearchContainer;
